refactor: Split course-attempts.component.tsx by modules
